+++ tests/nodes/test_ifnode.py
@@ -1,0 +1,7 @@
+from stilus.nodes.ifnode import If
+
+
+# TODO: implement complete test
+def test_property():
+    iff = If('condition')
+    assert iff.node_name == 'if'